simplify frontpage poll type
